convert json-api failure tests to a suite
