+++ ledger-service/http-json/BUILD.bazel
@@ -488,7 +488,70 @@ test_suite(
 )
 
 [
-    da_scala_test(
+    da_scala_library(
+        name = "failure-tests-lib-{}".format(edition),
+        srcs = glob(["src/failurelib/scala/**/*.scala"]),
+        plugins = [
+            kind_projector_plugin,
+        ],
+        scala_deps = [
+            "@maven//:com_typesafe_akka_akka_http_core",
+            "@maven//:com_typesafe_scala_logging_scala_logging",
+            "@maven//:io_spray_spray_json",
+            "@maven//:org_scalactic_scalactic",
+            "@maven//:org_scalatest_scalatest_core",
+            "@maven//:org_scalatest_scalatest_freespec",
+            "@maven//:org_scalatest_scalatest_matchers_core",
+            "@maven//:org_scalatest_scalatest_shouldmatchers",
+            "@maven//:org_scalaz_scalaz_core",
+            "@maven//:org_typelevel_cats_core",
+            "@maven//:org_typelevel_cats_effect",
+        ],
+        scalacopts = hj_scalacopts,
+        deps = [
+            ":http-json-{}".format(edition),
+            "//bazel_tools/runfiles:scala_runfiles",
+            "//daml-lf/data",
+            "//language-support/scala/bindings-akka",
+            "//ledger-api/rs-grpc-bridge",
+            "//ledger-api/testing-utils",
+            "//ledger-service/fetch-contracts",
+            "//ledger/ledger-configuration",
+            "//libs-scala/contextualized-logging",
+            "//ledger-service/http-json-cli:{}".format(edition),
+            "//ledger-service/http-json-testing:{}".format(edition),
+            "//ledger-service/jwt",
+            "//ledger-service/utils",
+            "//ledger/caching",
+            "//ledger/ledger-api-auth",
+            "//ledger/ledger-api-common",
+            "//ledger/ledger-resources",
+            "//ledger/metrics",
+            "//ledger/participant-integration-api",
+            "//ledger/participant-integration-api:participant-integration-api-tests-lib",
+            "//ledger/sandbox-common",
+            "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
+            "//ledger/sandbox-on-x:sandbox-classic-server",
+            "//libs-scala/db-utils",
+            "//libs-scala/ports",
+            "//libs-scala/ports:ports-testing",
+            "//libs-scala/postgresql-testing",
+            "//libs-scala/resources",
+            "//libs-scala/scala-utils",
+            "//libs-scala/timer-utils",
+            "@maven//:eu_rekawek_toxiproxy_toxiproxy_java_2_1_3",
+            "@maven//:io_dropwizard_metrics_metrics_core",
+            "@maven//:org_scalatest_scalatest_compatible",
+        ],
+    )
+    for edition in [
+        "ce",
+        "ee",
+    ]
+]
+
+[
+    da_scala_test_suite(
         name = "failure-tests-{}".format(edition),
         srcs = glob(["src/failure/scala/**/*.scala"]),
         data = [
@@ -519,6 +582,7 @@ test_suite(
         scalacopts = hj_scalacopts,
         deps = [
             ":http-json-{}".format(edition),
+            ":failure-tests-lib-{}".format(edition),
             "//bazel_tools/runfiles:scala_runfiles",
             "//daml-lf/data",
             "//language-support/scala/bindings-akka",
